+++ CHANGELOG.md
@@ -6,8 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## [X.X.X][X.X.X]
 
 ### Added
-- Allow EPS payment type without bank selection in shop.
-- Example for EPS without bank selection.
+- Parameter group to allow Online Transfer payments (e.g. EPS, Giropay and iDeal) without bank selection in shop.
+- Example for EPS and Giropay without bank selection.
 
 ## [v1.7.1][v1.7.1]
 
@@ -171,3 +171,5 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 [v1.6.0]: https://github.com/heidelpay/php-payment-api/compare/v1.5.0...v1.6.0
 [v1.6.1]: https://github.com/heidelpay/php-payment-api/compare/v1.6.0...v1.6.1
 [v1.6.2]: https://github.com/heidelpay/php-payment-api/compare/v1.6.1...v1.6.2
+[v1.7.0]: https://github.com/heidelpay/php-payment-api/compare/v1.6.2...v1.7.0
+[v1.7.1]: https://github.com/heidelpay/php-payment-api/compare/v1.7.0...v1.7.1